[SCRIP] Add admin_policy_enforced error in OAuth2AccessTokenFetcher
This cl adds logic to parse a new error string - admin_policy_enforced. This error is converted to
GoogleServiceAuthError::SCOPE_LIMITED_UNRECOVERABLE_ERROR.

Also this cl adds a subreason similar to InvalidGaiaCredentialsReason
for scope limited errors. This subreason will be used to disable sync
for enterprise policy for some reasons, but not for all.

Bug: b:352246509
Change-Id: Iabbfde58870a9a469eb540c207da2dbe391e42ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179438
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@google.com>
Reviewed-by: Monica Basta <msalama@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422446}