PA: Remove pool offset freelists feature
*   This was already launched in https://crrev.com/c/6106250.

*   It was default-enabled in V8 at https://crrev.com/c/6090745.

At any rate, this is now beyond the control of the variations framework.
If we hypothetically needed to land an emergency stoppage, we would do
so by directly editing the init sequence of `PartitionRoot`.

Since the `base::Feature` governing pool offset freelists is no longer
used (and not useful in any hypothetical), this CL trivially removes it.

This CL also incidentally default-enables pool offset freelists in
most uses of `PartitionOptions`. We maintain the seam in
`PartitionOptions` for the time being in case we need an emergency stop
in V8.

Note that this CL still leaves Chromium burdened with the freelist
dispatcher, which we will remove soon.

Bug: 40274683
Change-Id: Iddb8e5e4807675b4c5b3d618206e7957d14223d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6277911
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Sergei Glazunov <glazunov@google.com>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422468}