Update old canvas test generator to not add empty notes
The new version of the canvas test generator doesn't add a note tag if
the note is empty. Updating the old generator to do the same will
minimize the diff when we migrate the remaining tests from meta.yaml to
the new test generator.

Bug: 40207206
Change-Id: I902439f2ef2d5609a310661a6bbeb8aa9a88e49d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6204425
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422504}