Introduce CountryInfo
This will be used in order to represent country attributes. Using the
currently available `Address` class would be inconvenient because:

- This would allow callers to read/write info of unrelated types (e.g
  ZIP_CODE) from the country attribute, which is not pretty, unless we
  add additional filtering logic on AttributeInstance's side, which is
  not the preferred direction.

- This would conflict with other address types because the full value
  cannot be fetched in the same way from country attributes and other
  address attributes.

This CL therefore introduces a class that allows reading and writing
country values and provides a clear API that defines when to expect and
provide country names vs. codes.

Bug: 389625753
Change-Id: I94fa00061e159bc40f77369cae2cb6ee5e91972a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276195
Reviewed-by: Norge Vizcay <vizcay@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1422524}