[iOS] Auto-deletion IPH Display Logic
In this CL, the business logic that decides whether to display the
Auto-deletion IPH is implemented. The IPH is should be shown if the
user
1. The downloaded file is larger than 20MB OR
2. The device's storage has less than 2% of capacity available OR
3. The user has cleared browsing history in the last 60 days OR
4. The user has used incognito in the past 60 days OR
5. The user is actively downloading content in Incognito.

In a follow-up CL, the IPH's UI will be implemented. A screen
recording of this CL in action has been attached below.
https://drive.google.com/file/d/1SSwl3IAirRrkPJ4jyWWxnCWTlt5rBs_Q/view?usp=sharing&resourcekey=0-FIkHfekBTfl0XTIeW4CU3w

Bug: 390199773
Change-Id: Id538b6e49261b283f906d58aeb33e681367cff83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6282390
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422551}