[Android][Signin] Add sign-in flow default toast error
Currently, the sign-in flow shows an error toast only when the sign-in
is blocked by policies. This CL makes the flow show a similar toast for
the other error cases when sign-in can't start for consistency.

It also converts the existing
SigninAndHistorySyncActivityLauncherImplTest to unit test and adds
coverage for the error Toast UI using ShadowToast rather than checking
the view, to avoid potential flakiness as we had such issue in the past.

Last, the CL fixes some links in TODOs and make then point to open
issues.

Bug: 376251506
Change-Id: I0f77a6a42397d54ec1b6484cbc1486ff8b3b8fc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276527
Reviewed-by: Tanmoy Mollik <triploblastic@google.com>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1422555}