[Passkeys] Delay showing autofill popup without user interaction
For pages using WebAuthn Conditional UI, there is a race condition on
page load between password autofill becoming available and the WebAuthn
request being processed. Passkeys cannot be displayed in the autofill
UI until the request processing has reached a far enough point.

The problem becomes acute when the feature to enable the password
manager autofill popup due to field autofocus, without user
ineraction, is enabled. Popups will often appear with passkeys missing.

This change addresses that by adding a delay in some cases.
Specifically, if the popup is triggered due to autofocus, and it
has the `webauthn` tag, and the passkey list has not yet been
populated, it will set a timer for 4 seconds.

If the timer expires, the popup will be shown without passkeys.
If passkeys become available in the intervening time, then they
cause another "fill" to be triggered, which will result in the popup
appearing immediately with passkeys shown.
If the user clicks on the input field before the timer expires, then
the popup will be shown without passkeys (this is the same behaviour
as exists today on stable).

Bug: 326259918
Change-Id: I711dd5023caebd257dd47f1fa6990b4b67bcad48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6266488
Reviewed-by: Maria Kazinova <kazinova@google.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422579}