[Passkeys] Allow multiple callers to wait for conditional UI passkeys
Currently `ChromeWebAuthnCredentialsDelegate::RetrievePasskeys` allows
`PasswordFormManager` to get notified when passkeys become available for
display in autofill UI.

`PasswordAutofillManager` now also needs a notification if it is
waiting on passkeys before displaying an autofill popup that has
already been triggered.

This change allows `ChromeWebAuthnCredentialsDelegate` to keep a
vector of closures to invoke when passkeys are ready, rather than a
single callback.

It also renames `RetrievePasskeys` to
`RequestNotificationWhenPasskeysReady`. The existing name and
comments around it suggest that this triggers the actual retrieval,
which is incorrect. The passkey list is populated as a result of the
WebAuthn Javascript call. The method in question is just a way to
signal that a caller is waiting for that to finish.

Bug: 326259918
Change-Id: Ib3de7523bd3b7a3a3de87aa2872446039a592100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281052
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1422580}