[compare] Add product info batch api
This new api allows users to fetch all of the product info for a set
of urls at once. The behavior of this api is largely the same as its
single-url counterpart but will fetch data for any urls not already
on-device in a single batch to optimization guide's on-demand api.
The urls being requested still need to be referenced in the commerce
info cache to be allowed.

Bug: 391333469
Change-Id: I46f53f1454950b1dfb18dca3c89b70ab72345894
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6236866
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422595}