[Gap Decorations]: Remove Code Related to Row Gap Fragmentation
This CL removes code related to row gap fragmentation in the gap
geometry data structure. Previously, when a grid container is
fragmented, it was possible for only the start or end offsets of a row
gap to be present in a fragment. However, following the CSSWG's
resolution[1], gaps should remain suppressed if they fall on a
fragmentation boundary or if they are the last content in a
fragmentainer within a fragmentation context. Following this, for any
given row gap, we can now assume that the start and end offset will
always be present in that fragment.

[1]: https://github.com/w3c/csswg-drafts/issues/11520

Bug: 357648037
Change-Id: If90c2a01c628e3c22f8a4ea08e1637e745f94ab3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6255214
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1422602}