[AutofillAi] Controller exposes entities changes to be shown in UI
This CL introduces a method to the controller that will make it
possible for the UI to build the following two dialogs:

Update: https://screenshot.googleplex.com/6B2arDQkY9dvALS.png
Save: https://screenshot.googleplex.com/3HjtJzrDGmjGiT3.png

For the update case a follow Up CL will render the first N items
that are from the new entity and were either added or changed. It will
also add the tags depending on the update type.

Once this condition fails it will add a margin and render the rest of
the attribute changes. It will add a strikethrough line if the value is
not from the new entity.

For the save case, It will simply render everything. Note that a
IsSaveDialod or equivalent method will be added to the controller.

Bug: 389629676
Change-Id: I1eeb86d40d0719e7291c78c85a6ff12302a22aca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281092
Commit-Queue: Bruno Braga <brunobraga@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1422621}