[User Education] Accommodate async IPH API in webapp logic.
This fixes some logic that did not account for the fact that
MaybeShowFeaturePromo now reports its success or failure on a callback,
and in User Education 2.5 there may be a small delay between request
and actual show.

Now, the correct handling will happen when the promo should be dismissed
after it is shown.

Tests have also been updated to accommodate the async nature of showing
promos.

Bug: 369403281
Change-Id: Id62baf6dbe955f7ea9db16003dd4f4206f21cfa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6276968
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422640}