[mtb-cct] Per-embedder adaptive button customization
Adaptive button performs post-prediction filtering to take the device
form factor into account and rule out NewTab or bookmark button which
are already shown on the UI surface on Tablet. But this policy works
only for BrApp, not for CCT which requires a different policy that
will let it avoid duplication with the action buttons specified by
developers.

This CL makes it possible to inject a customized policy object
(AdaptiveToolbarBehavior) so that each embedder (Tabbed/CustomTab)
provide its own way to pick the right one from the list of buttons
predicted by segmentation results.

AdaptiveToolbarBehavior is the interface that allows per-embedder
customization for 3 things:

- Instantiate embedder-specific action buttons
- Filter the segmentation results to pick the button
    to display (explained above)
- Whether the raw segmentation result may be used

|AdaptiveToolbarUiCoordinator| is a new class introduced in this CL.
|RootUiCoordinator|(RUC) delegates most MTB-related tasks to it.
It works as a bridge between RUC and |AdaptiveToolbarButtonController|.

|Tabbed/BaseCustomTabRootUiCoordinator| implements its own
 |AdaptiveToolbarBehavior| objects and passes it to
 |AdaptiveToolbarButtonPredictor| to apply the customized filter.

This CL also deletes |AdaptiveToolbarBridge.getSessionVariantButton|
as it was left unused after the new, general method
|getSessionVariantButtons| (returns multiple results) was added.

BYPASS_LARGE_CHANGE_WARNING: Simple move of ~400 LOC verbatim from RootUiCoordinator to the new class was detected as a large change

Bug: 391931152
Change-Id: I8a5692596a54f6a38f5e62d9bbb1ee1e502fdcea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6245676
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Salvador Guerrero Ramos <salg@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422674}