[sync] Add test-only flag to simulate ping-pong scenarios
If the flag is enabled, meant for manual tests only, it causes clients
to ensure that BookmarkSpecifics.unique_position is cleared. Any remote
updates (initial download or incremental) that has this field populated
will be immediately corrected with a commit that clears this field. This
is the exact opposite of what the default behavior is, which tries to
ensure that this field is always populated.

Therefore, two clients (or more) with a different value set for this
flag will enter a ping-pong fight, one of them trying to clear the field
and the other one trying to populate it. Because bookmarks have a local
nudge delay of 2 seconds, it is expected (and verified empirically) that
each client will cause a commit traffic of 0.5 QPS under good network
conditions, until the account gets throttled (after 20 minutes or so).
This QPS can be higher if the account has many bookmarks, which means
reuploading all bookmarks would require more than one commit request.

Change-Id: Ie8e5aed8564c442251de94e35c3ba69cc488578b
Bug: 390112136
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6285889
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1422680}