[Autofill] Pass EntityInstance around instead of flat_map
This CL lets FillOrPreviewForm() take an `EntityInstance*` instead
of a `base::flat_map<FieldGlobalId, std::u16string>` as part of the
FillingPayload variant. This change is consistent with how it works
for addresses and credit cards.

This propagates to the `Suggestion` payload, which is now a
`base::Uuid` instead of a `base::flat_map<FieldGlobalId,
std::u16string>`.

Bug: 397505149
Change-Id: I9108a20ad8246f571559281f7b62bf0e2480faac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6282006
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1422714}