[ios blink] Allow net and url deps in extensions
The credential_provider_extension depends on passkey_model
which then depends on components/keyed_service/core. This
dependency is just for a single file for the parent of
the PasskeyModel which is KeyedService. This relationship
could be fixed if keyed_service.h was placed in a source_set
that just contained the header. However a further problem
exists in that passkey_model_utils.h end up using device/fido
and device/bluetooth and these then end up depending on
tracing. Tracing then depends on //net and //url. So the
most comprehesive approach is to allow //net and //url since
trying to avoid tracing in an extension is probably near
impossible.

Bug: 356196908
Change-Id: Ie24ef945ff1398fd39cc5636141c6da7f4045585
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288106
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422728}