[EducationalTip] Fix bug where the card's title gets truncated.
This CL resolves a bug where the title of the educational tip quick
delete promo card gets truncated on tablets. It updates the title to fit
across two lines and adjusts the description to be just one line.
Conversely, if the title fits within a single line, the description
should span two lines and the title will return to one line.
The video that shows this change:
https://drive.google.com/file/d/19CL4WfsMUeu_kmy65LsUxItKf36krpJp/view?usp=sharing&resourcekey=0-ya7MxWnnY5JsW4drwcmMYw

Bug: 343231383
Change-Id: Ib4860dae81e616c5751646b005a0d435b053253e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6256778
Commit-Queue: Xinyi Ji <xinyiji@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422780}