android: Rework class and target stats script
Re-order the list of targets that are output by the script in the order
that makes it easiest to be null-annotated. Targets that have all of
their dependencies already annotated are the easiest as all the changes
are within that target. These can also be annotated in parallel.

By default only targets that have all their dependencies already
annotated are shown, --all can be used to display the rest as well.

Add a --verbose flag in case more output is desired. Make the default
the easiest to view and copy/paste/use.

Bug: None
Change-Id: I288015d628af4ab32704b55e4b664b062ea67f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6285691
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422790}