gpu: win: Acquire DComp texture available fence immediately on retiring
This CL changes moves the DComp texture available fence acquisition from
immediately before the subsequent write to immediately after retiring
the texture.

This change moves the possible case of failure to acquire the available
fence to the frame with a problematic DComp commit, rather than a later
rasterization task.

This is valid because DComp commit is not an event that affects the
validity of the available fence, as previously thought. As long as the
DComp texture is not placed back into any visual tree and we do not
release the DComp texture, the available fence remains valid.

Change-Id: I96be19a6271eda6d9f5060baf837059909b63204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6271326
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1422796}