Reduce the weight of mojo_binding_context.h
This file accounts for about 838 MB (0.2%) of Chrome's expanded source
size, but it is much heavier than it needs to be.

Move the MojoJS specific logic (which is rarely used anyway) into a
supplement created on-demand, allowing the header to expose only the
signatures. Fix up missing includes in a number of source files.

Bug: 40318405
Change-Id: I8c78ae52674ac3e142362740cfb6f7b516f36f3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6284085
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422825}