Add empty preference screen for "Appearance" settings.
This CL adds a new empty preference screen for "Appearance" settings.
The screen will be populated in a later CL and its presence is gated by
the AndroidAppearanceScreen feature flag. When the "Appearance" menu
option is visible in main settings, "Theme" and "Toolbar shortcut" are
removed as they will be moved to the new screen in later CLs.

Design: http://shortn/_oxzSbwNuTe

Bug: 397775252
Change-Id: I90df378241173d06489322fd901e2237f1c040ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288038
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422853}