[SxS] Add (non-functional) resize handle
Creates a MultiContentsResizeArea class which is visually similar to the
side panel's resize area. This can respond to left/right keyboard events
to resize as well as mouse click+drag, and has a keyboard focusable
icon. Currently emits events to a stubbed out OnResize method, will
implement this in a follow-up to actually resize the WebContents as
relative to each other.

Bug: 393450761
Change-Id: Ic8945f0e25995cfb5a9c19bd9d4e097050f5b95c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281806
Commit-Queue: Emily Shack <emshack@chromium.org>
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422859}