Make HttpStreamParser::SeekableIOBuffer use SetSpan().
IOBuffer's internal `data_` and `size_` fields needs to be replaced with
a single span field, to help protect against OOB reads/writes. To do
this, IOBuffer subclasses need to be switched to using SetSpan(), and
never access those fields directly.

This CL updates SeekableIOBuffer to use SetSpan() instead of updating
those fields directly. It also replaces its raw_ptr field pointing at
the start of the buffer with a raw_span, renames its `size_` (which was
different from IOBuffer::size_) to `bytes_written_`, renames the
ambiguously named `used_` field to `bytes_consumed_`, and removes the
`capacity_` field (which is the same as the size of the newly
introduced raw_span field).

Bug: 396621713
Change-Id: Icdb094d2a34b459ca7d730afc021e8318d5c7412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6286888
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422874}