Spanify HttpConnection::ReadIOBuffer
ReadIOBuffer provided StartOfBuffer() and GetSize() methods, which effectively returned a span. This replaces them with a single
readable_bytes() method that returns a span. ReadIOBuffer wraps a
GrowableIOBuffer, and readable_bytes() is logically the same as
GrowableIOBuffers::span_before_offset(), so the new method calls
that. The new method is not called span_before_offset() because
focusing on what's readable is more consistent with other method
ReadIOBuffer documentation and method names.

Using a span adds safety checks not present on raw buffers, and this
simplifies the code a bit as well. Also, ReadableIOBuffers had
GetSize() and size() methods, which return sizes of different halves
of the managed buffer, so removing GetSize() will hopefully reduce
confusion.

Bug: 396621713, 40284755
Change-Id: I1d9c769f2c2e3e3892982d0c002e7fd1f092c08d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6282326
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422880}